+++ FileConfig.js
@@ -42,13 +42,8 @@ export const FileConfig = {
 	},
 
 	toDistPath(path) {
-		if (path.startsWith("frontend/")) {
-			// is this needed? attempt to remove replaceAll on windows and see if it still works
-			const p = path.replaceAll("\\", "/").replace("frontend/", "")
-			return `${FileConfig.dist}/${p}`
-		}
-		else {
-			throw new Error(path + " does not start with frontend/")
-		}
+		// is this needed? attempt to remove replaceAll on windows and see if it still works
+		const p = path.replaceAll("\\", "/").replace("frontend/", "")
+		return `${FileConfig.dist}/${p}`
 	},
 }
