+++ frontend/client/shared/SuperClass.js
